Python setup script to create GoogleSpanner DB
setupGoogleSpanner.py will create the GoogleSpanener instance and/or
datbase if they do not already exist.
google-cloud-spanner is required. Run "pip3 install -r requirements.txt"
to install the module. pip3 is required because this is a python3 script.
So run it with pyrhon3 too. May also work with python2.7 though, I'm
haven't tested.
The GCLOUD_PROJECT and GOOGLE_APPLICATION_CREDENTIALS environment
variables need to be set before running the script. Read the comments at
the top of the script for more information on what these should be.
The config options for the script are currently hardcoded. This is
probably fine, they'll just need to be modified by the user.
Feel free to modify the script however you want.